# src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "BookStore APIs exposed from a Java EE back-end to an Angular front-end",
    "version" : "1.0.0",
    "title" : "BookStore APIs"
  },
  "host" : "localhost:8080",
  "basePath" : "/bookstore-back/api",
  "tags" : [ {
    "name" : "Book"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/books" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns all the books",
        "description" : "",
        "operationId" : "getBooks",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Books found",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Book"
              }
            }
          },
          "204" : {
            "description" : "No books found"
          }
        }
      },
      "post" : {
        "tags" : [ "Book" ],
        "summary" : "Creates a book given a JSon Book representation",
        "description" : "",
        "operationId" : "createBook",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Book to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Book"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "The book is created"
          },
          "415" : {
            "description" : "Format is not JSon"
          }
        }
      }
    },
    "/books/count" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns the number of books",
        "description" : "",
        "operationId" : "countBooks",
        "produces" : [ "text/plain" ],
        "responses" : {
          "200" : {
            "description" : "Number of books found",
            "schema" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "204" : {
            "description" : "No books found"
          }
        }
      }
    },
    "/books/{id}" : {
      "get" : {
        "tags" : [ "Book" ],
        "summary" : "Returns a book given an id",
        "description" : "",
        "operationId" : "getBook",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Book found",
            "schema" : {
              "$ref" : "#/definitions/Book"
            }
          },
          "400" : {
            "description" : "Invalid input. Id cannot be lower than 1"
          },
          "404" : {
            "description" : "Book not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Book" ],
        "summary" : "Deletes a book given an id",
        "description" : "",
        "operationId" : "deleteBook",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "minimum" : 1,
          "pattern" : "\\d+",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Book has been deleted"
          },
          "400" : {
            "description" : "Invalid input. Id cannot be lower than 1"
          },
          "500" : {
            "description" : "Book not found"
          }
        }
      }
    }
  },
  "definitions" : {
    "Book" : {
      "type" : "object",
      "required" : [ "isbn", "title" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Identifier"
        },
        "title" : {
          "type" : "string",
          "description" : "Title of the book",
          "minLength" : 1,
          "maxLength" : 200
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the book",
          "minLength" : 1,
          "maxLength" : 10000
        },
        "unitCost" : {
          "type" : "number",
          "format" : "double",
          "minimum" : 1
        },
        "isbn" : {
          "type" : "string",
          "minLength" : 1,
          "maxLength" : 50
        },
        "publicationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "nbOfPages" : {
          "type" : "integer",
          "format" : "int32"
        },
        "imageUrl" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string",
          "enum" : [ "ENGLISH", "BENGALI", "HINDI" ]
        }
      },
      "description" : "Book resource representation"
    }
  }
}